requirements.txt
fastapi==0.111.0
pinecone-client==5.0.0
sqlalchemy~=2.0
alembic-postgresql-enum==1.1.0
alembic~=1.13
loguru==0.7.2
python-dotenv
pydantic-settings
pyjwt
loguru
asyncpg
psycopg2
aiogram
redis
aioredis==2.0.1